requirements.txt
flask
gunicorn
pandas
azure-storage-blob
